+++ senders/types.go
@@ -5,37 +5,19 @@ import (
 	"github.com/wavefronthq/wavefront-sdk-go/histogram"
 )
 
-type SpanTag struct {
-	Key   string
-	Value string
-}
-
-type SpanLog struct {
-	Timestamp int64             `json:"timestamp"`
-	Fields    map[string]string `json:"fields"`
-}
-
-// SpanLogs is for internal use only.
-type SpanLogs struct {
-	TraceId string    `json:"traceId"`
-	SpanId  string    `json:"spanId"`
-	Logs    []SpanLog `json:"logs"`
-	Span    string    `json:"span"`
-}
-
 // MetricSender Interface for sending metrics to Wavefront
 type MetricSender interface {
-	// Sends a single metric to Wavefront with optional timestamp and tags.
+	// SendMetric sends a single metric to Wavefront with optional timestamp and tags.
 	SendMetric(name string, value float64, ts int64, source string, tags map[string]string) error
 
-	// Sends a delta counter (counter aggregated at the Wavefront service) to Wavefront.
+	// SendDeltaCounter sends a delta counter (counter aggregated at the Wavefront service) to Wavefront.
 	// the timestamp for a delta counter is assigned at the server side.
 	SendDeltaCounter(name string, value float64, source string, tags map[string]string) error
 }
 
 // DistributionSender Interface for sending distributions to Wavefront
 type DistributionSender interface {
-	// Sends a distribution of metrics to Wavefront with optional timestamp and tags.
+	// SendDistribution sends a distribution of metrics to Wavefront with optional timestamp and tags.
 	// Each centroid is a 2-dimensional entity with the first dimension the mean value
 	// and the second dimension the count of points in the centroid.
 	// The granularity informs the set of intervals (minute, hour, and/or day) by which the
@@ -45,7 +27,7 @@ type DistributionSender interface {
 
 // SpanSender Interface for sending tracing spans to Wavefront
 type SpanSender interface {
-	// Sends a tracing span to Wavefront.
+	// SendSpan sends a tracing span to Wavefront.
 	// traceId, spanId, parentIds and preceding spanIds are expected to be UUID strings.
 	// parents and preceding spans can be empty for a root span.
 	// span tag keys can be repeated (example: "user"="foo" and "user"="bar")
@@ -55,6 +37,16 @@ type SpanSender interface {
 
 // EventSender Interface for sending events to Wavefront. NOT yet supported.
 type EventSender interface {
-	// Sends an event to Wavefront with optional tags
+	// SendEvent sends an event to Wavefront with optional tags
 	SendEvent(name string, startMillis, endMillis int64, source string, tags map[string]string, setters ...event.Option) error
+}
+
+type SpanTag struct {
+	Key   string
+	Value string
+}
+
+type SpanLog struct {
+	Timestamp int64
+	Fields    map[string]string
 }
